+++ wolfsentry/wolfsentry_errcodes.h
@@ -48,15 +48,15 @@ typedef int32_t wolfsentry_errcode_t;
 
 #if defined(__GNUC__) && !defined(__STRICT_ANSI__)
 #define WOLFSENTRY_ERROR_ENCODE_1(x) ({                                      \
-    wolfsentry_static_assert(((x) >= -WOLFSENTRY_ERROR_ID_MAX)               \
+    wolfsentry_static_assert2(((x) >= -WOLFSENTRY_ERROR_ID_MAX)              \
                    && ((x) <= WOLFSENTRY_ERROR_ID_MAX),                      \
                   "error code must be -"                                     \
                   _q(WOLFSENTRY_ERROR_ID_MAX)                                \
                   " <= e <= "                                                \
                   _q(WOLFSENTRY_ERROR_ID_MAX) )                              \
-    wolfsentry_static_assert(__LINE__ <= WOLFSENTRY_LINE_NUMBER_MAX,         \
+    wolfsentry_static_assert2(__LINE__ <= WOLFSENTRY_LINE_NUMBER_MAX,        \
                   "line number must be 1-" _q(WOLFSENTRY_LINE_NUMBER_MAX) )  \
-    wolfsentry_static_assert((WOLFSENTRY_SOURCE_ID >= 0)                     \
+    wolfsentry_static_assert2((WOLFSENTRY_SOURCE_ID >= 0)                    \
                   && (WOLFSENTRY_SOURCE_ID <= 0x7f),                         \
                   "source file ID must be 0-" _q(WOLFSENTRY_SOURCE_ID_MAX) ) \
     WOLFSENTRY_ERROR_ENCODE_0(x);                                            \